Move function to utils extract configuration from main start using global variables
